Forward the access token to the CKAN API instead of a configured API key
This behaviour can be overridden by configuring the adapter's 'use_apikey' option: setting it to True will assume that the ODP API was called with a CKAN API key in the Authorization header (rather than an access token), and will forward the API key appropriately. In order for this to work, normal access token validation will have to be disabled by also setting the application's 'security.no_access_token_validation' option to True.

Changed the way the CKAN URL is configured: it may be set in the ODP API's config.yml, but optionally overridden by the CKAN_URL environment variable (e.g. for Docker deployments).