Define simpler, broader OAuth2 scopes
OAuth2 scopes != permissions

Regardless of the fact that many APIs (including e.g. Google APIs) define distinct scopes for reading, writing and/or managing various type of resource, the purpose of an OAuth2 scope is to limit what an app can do on behalf of a user, *not* to specify actual permissions. Permissions must still be evaluated on the resource server, regardless of what the access token indicates, and typically vary per user per role per actual resource *instance*. To fully specify permissions in an access token, we would need to issue distinct access tokens for every resource, which is totally impractical.

So, sticking to the intended purpose of OAuth2 scopes, we need only define a scope to represent a broad functional area of the ODP, or a broad type of resource. We can thus adequately constrain a client app to only making API requests that are suited to its purpose.