:robot: Authorize each request automatically
Reworked dependency injection so that we put config, adapter and auth data onto the request state for each request, before the request gets to the route function. This (a) reduces boilerplating of route function signatures, and (b) enables configuration of routers. From (b) we fix the way routers and adapters are hooked up - instead of configuring an adapter to handle routes (messy!), we configure a router to use an adapter (neat!).

This change also switches from directly calling the Hydra admin API for token validation and introspection to calling the intermediary ODP Accounts API, which handles the decoding of privileges info from access tokens.