Simplify error response model
For CKAN validation errors, just represent the error dict as a string. This means we are using the same response model {"detail": "err"} across the board. Putting the CKAN dict into the error response complicates things for clients of the ODP API, and is impossible to accurately document anyway. The CKAN API is a black box and there's no point in trying to expose its innards to the world. We still get all the error info out in the response, which has to be interpreted by a developer anyway; whether it's a string or a dict makes no difference.