Reorganise the project
Remove the runnable main file, the explicit dependency on uvicorn, and unnecessary related config (host and port). These concerns are externalised; guidance is provided on running the API in development in the readme.
Add a PATH_PREFIX option to enable the Swagger docs to work when mounting the API at a URL path.