Remove client login capability of the identity service
This change removes the identity service homepage and OAuth2 client dependencies and, therefore, the ability to login to the identity service. It keeps the service focused on its real job - which is to provide login, signup, etc, for other applications.

When we get to developing the identity management functions - e.g. update profile, request/assign role, etc - these should either be incorporated into the login flow, or done in a separate app.