Create a unified development environment
Use the develop directory as a combined repository for config and readme info for all services. Instead of a plethora of different runtime, config and readme locations, we'll have just one. Thus far only applied to the ODP admin and public APIs - which also allowed for simplification of the API module structure.