Rework publishing framework
This change sets up the metadata_status table as the single source of truth for ODP metadata - in effect, the ODP catalogue - for systems such as the Elasticsearch-based metadata discovery catalogue, and any external entity that wishes to harvest metadata from the ODP. The table is populated from the internal metadata management system; records are only added if they've been published in the metadata manager, although we retain records that are later un-published to enable external harvesters to delete as applicable.

Catalogue publishing is now made *less* generic - it makes more sense to have all the logic for determining what and how to sync with a particular catalogue in the catalogue class itself, instead of having contrived and difficult-to-read code in a parent class, catering for as-yet-unimplemented use cases.

Also, this closes #4