Exclude records without DOIs from DataCite sync selection
Because a DOI, once set, is permanently pinned to a record, we know that if a record does not have a DOI, it could never have previously had a DOI and therefore could never have previously been published to DataCite. We can therefore exclude updated catalogue records without DOIs from being selected as candidates for synchronization (publishing/unpublishing) with DataCite.

Closes #6