Do not cascade delete to child institutions
Cascading a delete of an institution to child institutions will fail if any child (or indeed the institution itself) has member users. However, this does not cater for the case of institutions with external dependents - e.g. metadata collections in the metadata management system. If the delete of a parent institution cascades to a child without members but with external collections, those collections will be orphaned. So instead, we simply break the parent-child link when deleting a parent.