Remove profile completion from the signup flow
And restore the previous quick login after verifying the user's email address.

The profile completion form - which has a 'first name' and a 'last name' field along with a login button - can easily be mistaken for a login form. In fact, I fell for this myself, entering my email in the first name field and my password in the last name field! This presents a security risk: we end up inadvertently storing plain-text passwords of any users who happen to mistake the profile form for a login form.