Enable view access to the admin app
Previously only an admin (or superuser) could read or write anything in the app. This was partly because it is nigh on impossible to dynamically change the visibility of row actions.

Refactored the view access logic to allow read access for anyone with an admin-scoped role, and hook into any attempt to perform an action and make sure the user has sufficient permission to make changes.