Improve access control implementation in the admin app
We do away with the notion of a superuser (only in this app; it can be gradually deprecated elsewhere), instead differentiating manager- and admin-level access: managers can manage institutions, users and privileges (though we prevent them from being able to grant admin rights); admins can do the same as well as modify system config (roles and scopes) and grant admin rights.

We also do away with config values identifying the admin role and admin scope - these are already enums, so the config is unnecessary.