Switch to an explicit Session registry usage pattern
Rename `session` to `Session`, since this variable represents a session registry/factory rather than a session instance - even though it behaves like a session instance.

Remove the `transaction` context manager. This affects (a) web requests: we will commit the transaction at the end of a POST/PUT/DELETE request, instead of doing multiple commits over the course of a single request; and (b) ad-hoc transactions in scripts (e.g. the metadata publishers), for which we can instead use the `Session.begin` context manager.