Use string id as primary key across the board :key:
We have already done this for config entities like role, scope, client, etc, an approach that was inspired by the use of meaningful string ids for OAuth2 clients. Now we extend this approach to project, provider, collection and metadata_schema, doing away with the unique 'key' column.

The idea with separating id and key columns was that users might want to change the key on the odd (rare) occasion, so it should be unique but not primary. But even changing such a key can have deleterious consequences: there can be many external references to it, in catalogues and client scripts, for example. The UI would have to warn users against changing the key and/or make it difficult to do so.

The new approach says that when you create a project or any other object (besides record and user), you specify a suitable and meaningful id, and that's it, you can't change it.

Amongst the advantages of this approach, table joins are often not necessary since associations are immediately obvious, and foreign keys can be used directly for URL generation in the UI.