Drop client-specific roles
As with project-specific roles, there's no fundamental use case for this. We have a handful of internally-developed client apps and a handful of developers - and it's fair to say that we can and should trust each other not to tinker with each other's client configurations.

If any need arises for external app/script access, for the time being it makes sense to create the client config on behalf of the external developer.