Revise data model
Define role and client similarly: role represents a set of permissions associated with a user, client a set of permissions associated with an application. Both may optionally be linked with a provider, which limits the applicability of any attached provider, collection and record scopes. Aligning role and client in this way makes it clearer how the effective set of permissions for a user logged in to a client is determined.