Solve an interesting edge case that threw a rare false negative
API list methods by default return a list of objects ordered by id, using a SQL order by clause. In the tests, we compare this returned list with the list of factory-generated objects which we sort using a standard Python list sort on the object ids.

However, given two sample ids such as 'Reactive' and 'Re-contextualized' (produced by the Faker library), PostgreSQL will sort them as:

Reactive
Re-contextualized

PostgreSQL ignores hyphens in text comparisons. In contrast, Python will sort these values as:

Re-contextualized
Reactive

So even if the returned API results are correct, if two such items happen to be in the factory-generated batch for an API list test, the test will fail. We solve this by doing a Python sort on the result list returned from the API.