Revert "Support multiple allowed scopes for API routes"
This reverts commit c622fe005d9ed7b733d9c1649e6c63c92deafd0e.

Multiple allowed scopes per API route is not entirely compatible with OAuth2.
The problem was exposed by b9601de537e925e08ea80a2a5498d0d459d023b7, which
reduces the set of scopes that appear in a user's access token from those
requested by the client to those actually assigned to the user, even though
the effective authorization result within the API is unchanged.

Let's say we define the GET /record/ route as requiring ANY OF the
odp.record:read and odp.record:admin scopes. Now we'd expect a user with
the odp.record:admin scope (but without the odp.record:read scope) to be
able to access that route. However, Hydra reports the token as being
invalid, because it requires ALL OF the specified scopes for that route
to be associated with the access token.

Requiring a single scope per API route reduces the complexity of API
authorization. It also requires user roles to be explicitly configured
with scopes for each required route, eliminating the need for implicit
knowledge that, for example, admin scopes grant both read and write access.