Support client (external) and standard OAuth2 scopes in the data model
This is more sensible than trying to be clever in the API with distinguishing ODP system scopes from other kinds of scopes. This way, the database underpins all API and UI scope assignments. We can fix the set of standard scopes that are available (as we already do with ODP system scopes), and allow developers to set up custom scopes for client end use.