Change record PUT to be a create-or-update operation
And make it admin-scoped (just a scope rename, but aligns with other admin scope usages). This allows relevant roles (e.g. curators), as well as system processes (in particular, v1-to-v2 migration, but conceivably harvesters, etc), to set the record id for new records and transparently create/update for reentrant bulk operations.