Remove flags (merge with tags)
All the flag code is almost identical to all the tag code. Flags are really just a specialization of tags with a cardinality of 1 wrt records/collections. Thus, we can cater for them simply by introducing a `multi` (false for flags) attribute on the tag model.

This change simplifies application usage of flags/tags (including in publishing) since there is now just one list of tags on each of the record and collection models.