Support changing project, provider and collection ids
Such changes are rare but not unlikely - names of government departments change from time to time, for example. If such a change is required, it can now be done with an update directly to the relevant DB table. There is no need (and it would be risky) to support this in the API/UI, however.