Add client-collection link
It makes sense to keep object-specific client restrictions (as with roles, change from provider- to collection-specific).

This is then a boilerplate change across the board. We already use the client-object link extensively in API testing, which gives auth coverage that would be much more difficult using role-object links (because we don't have a logged in user in testing).

Also, there are use cases for collection-specific clients: the Obs DB harvester(s), writing only to Obs DB collection(s); and in future external clients for managing MIMS collections.