Remove provider_id and project_ids from API record models
This avoids having to account for changes to collection.provider or collection.projects that may have occurred subsequently to a record being published. Instead of publishing a record with a static snapshot of these collection properties, we can allow catalog clients read access to collections, projects and providers as appropriate, so they can look up those properties given the collection_id for a published record.