Prevent deletion of users who've created tag instances
At this stage we do not have a policy for what to do with tag instances created by a user when that user is deleted - for example if a staff member leaves, or a user requests their data to be deleted wrt POPI.

Cascading user deletions to tag instances might have unintended consequences, such as causing records to be un-published - besides the loss of curation data. For now, the safer option in terms of data integrity is to block such a deletion.