Adapt framework for publishing to DataCite
The form of the record committed to the published_record column of catalog_record is now unconstrained. For the SAEON catalog, we'll commit a structured record including extended info such as collection ids and tags, as well as multiple outputs for different schemas as applicable. In contrast for DataCite, we'll commit the exact form of the record that we'll post to DataCite; so the DataCite catalog in the ODP will effectively mirror the ODP holdings on DataCite servers.