Flag fields as required in API audit models
We are now saving all fields to the respective audit log not only for inserts and updates, but for deletes too - even though the values will be identical to the previous (insert/update) audit record for a given object. This means that if we want to see exactly what was deleted, we don't have to go searching for that previous audit record.