Factor out an ODP UI library
This is a starting point. The goal is to be able to deploy ODP UI apps independently of ODP core systems - in particular, with no DB access.

The prototypical use case is the MIMS catalogue, which will run on DFFE servers. It will integrate with the ODP identity service and ODP API.